requirements.txt
appconfigpy==0.4.2
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
DataProperty==0.43.1
decorator==4.4.0
Django==2.2.1
django-filter==2.1.0
django-haystack==2.8.1
django-multiselectfield==0.1.8
djangorestframework==3.9.4
drf-haystack==1.8.5
idna==2.8
importlib-metadata==0.15
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.4.0
Logbook==1.4.3
Markdown==2.6.11
mbstrdecoder==0.8.1
msgfy==0.0.7
nbformat==4.4.0
path.py==12.0.1
pathvalidate==0.28.2
pyrsistent==0.15.2
pytablereader==0.26.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
retryrequests==0.0.3
simplejson==3.16.0
SimpleSQLite==0.45.2
six==1.12.0
soupsieve==1.9.1
sqlitebiter==0.28.0
sqliteschema==0.16.2
sqlparse==0.3.0
tabledata==0.9.1
traitlets==4.3.2
typepy==0.6.0
urllib3==1.25.3
xlrd==1.2.0
zipp==0.5.1
